centralizing all plotting functions into a single file
